✨ Completed parsing header for Corewar

+++ lib/common/include/header/defs.h
@@ -15,5 +15,5 @@
 typedef struct s_header {
     char name[HEADER_NAME + 1];             // Name of the champion
     char comment[HEADER_COMMENT + 1];       // Comment of the champion
-    int body_size;                          // N bytes in the body
+    long int body_size;                          // N bytes in the body
 } header_t;
